Delete files while running, fix a million bugs, import actual data

+++ java/org/trailcatalog/importers/pipeline/collections/Mmap.kt
@@ -7,13 +7,21 @@ import java.io.File
 import java.io.RandomAccessFile
 import java.nio.channels.FileChannel
 import java.nio.channels.FileChannel.MapMode
+import java.util.concurrent.atomic.AtomicInteger
 
 open class MmapPList<T>(
     private val maps: List<EncodedInputStream>,
     private val serializer: Serializer<T>,
-    private val size: Long) : PList<T> {
+    private val size: Long,
+    private val fileReference: FileReference,
+) : PList<T> {
 
   private var shard = 0
+
+  override fun close() {
+    maps.forEach { it.close() }
+    fileReference.close()
+  }
 
   override fun estimatedByteSize(): Long {
     return size
@@ -34,6 +42,10 @@ open class MmapPList<T>(
 
 open class MmapPSortedList<T>(private val list: MmapPList<T>) : PSortedList<T> {
 
+  override fun close() {
+    list.close()
+  }
+
   override fun estimatedByteSize(): Long {
     return list.estimatedByteSize()
   }
@@ -51,7 +63,9 @@ open class MmapPSortedList<T>(private val list: MmapPList<T>) : PSortedList<T> {
   }
 }
 
-fun <T : Any> createMmapPList(type: TypeToken<out T>, fn: (Emitter<T>) -> Unit): () -> MmapPList<T> {
+fun <T : Any> createMmapPList(
+    type: TypeToken<out T>, handles: AtomicInteger, fn: (Emitter<T>) -> Unit): () -> MmapPList<T> {
+  if (handles.get() == 0) throw RuntimeException("${type}")
   val serializer = getSerializer(type)
   val file = File.createTempFile(cleanFilename("mmap-list-${type}"), null)
   file.deleteOnExit()
@@ -75,17 +89,21 @@ fun <T : Any> createMmapPList(type: TypeToken<out T>, fn: (Emitter<T>) -> Unit):
   println("PList (mmap) ${type} size ${size}")
 
   val fileChannel = FileChannel.open(file.toPath())
-  val maps = shards.map { s ->
-    EncodedInputStream(fileChannel.map(MapMode.READ_ONLY, s.start, s.length))
-  }
+  val fileReference = FileReference(file, handles)
 
   return {
-    MmapPList(maps, serializer, size)
+    val maps = shards.map { s ->
+      EncodedInputStream(fileChannel.map(MapMode.READ_ONLY, s.start, s.length))
+    }
+    MmapPList(maps, serializer, size, fileReference)
   }
 }
 
-fun <T : Comparable<T>> createMmapPSortedList(type: TypeToken<out T>, fn: (Emitter<T>) -> Unit):
-    () -> MmapPSortedList<T> {
+fun <T : Comparable<T>> createMmapPSortedList(
+    type: TypeToken<out T>,
+    handles: AtomicInteger,
+    fn: (Emitter<T>) -> Unit,
+): () -> MmapPSortedList<T> {
   val interceptedFn = { emitter: Emitter<T> ->
     var last: T? = null
     val checkingEmitter = object : Emitter<T> {
@@ -101,7 +119,7 @@ fun <T : Comparable<T>> createMmapPSortedList(type: TypeToken<out T>, fn: (Emitt
     fn(checkingEmitter)
   }
   return {
-    MmapPSortedList(createMmapPList(type, interceptedFn).invoke())
+    MmapPSortedList(createMmapPList(type, handles, interceptedFn).invoke())
   }
 }
